+++ src/core/lookup.js
@@ -1,39 +1,36 @@
-const lookup = async (page, lookupService = "https://api.ipify.org?format=json", isJSON = true, timeout = 30000) => {
+const lookup = async (page, lookupService = "https://api64.ipify.org?format=json", isJSON = true, timeout = 30000) => {
     const doLookup = async () => {
         return await page.evaluate((lookupService, timeout, isJSON) => {
             return new Promise((resolve) => {
-                const req = new XMLHttpRequest();
-                req.timeout = timeout;
-                req.onload = () => {
-                    if (req.status >= 200 && req.status <= 299) {
-                        resolve(isJSON ? JSON.parse(req.responseText) : req.responseText);
-                    } else {
-                        resolve(onLookupFailed(`Request from ${window.location.href} to ${lookupService} failed with status code ${req.status}`));
-                    }
+                const request = new XMLHttpRequest();
+                request.timeout = timeout;
+                request.onload = () => {
+                    if (request.status >= 200 && request.status <= 299) {
+                        resolve(isJSON ? JSON.parse(request.responseText) : request.responseText);
+                    } else {resolve(onLookupFailed(
+                        `Request from ${window.location.href} to ` + 
+                        `${lookupService} failed with status code ${request.status}`
+                    ))}
                 };
-                req.ontimeout = (error) => {
-                    resolve(onLookupFailed(`Request from ${window.location.href} to ${lookupService} timed out -> ${req.timeout} ms`));
-                };
-                req.open("GET", lookupService, true);
-                req.send();
+                request.ontimeout = (error) => {resolve(onLookupFailed(
+                    `Request from ${window.location.href} to ` +
+                    `${lookupService} timed out at ${request.timeout} ms`
+                ))};
+                request.open("GET", lookupService, true);
+                request.send();
             });
         }, lookupService, timeout, isJSON);
     };
     try {
         await page.setBypassCSP(true);
-        const functionName = "onLookupFailed";
+        const functionName = "$ppp_on_lookup_failed";
         if (!page._pageBindings.has(functionName)) {
-            await page.exposeFunction(functionName, (reason) => {
-                console.error(reason);
-                return;
+            await page.exposeFunction(functionName, (failReason) => {
+                console.error(failReason); return;
             });
         }
         return await doLookup();
-    } catch(error) {
-        if (error.message.startsWith("Execution context was destroyed")) {
-            return await doLookup();
-        }
-    }
+    } catch(error) {console.error(error)}
 };
 
 module.exports = lookup;